# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solvent API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/pairs/history": {
      "get": {
        "tags": [
          "crate::http::app::pairs"
        ],
        "operationId": "pair_history",
        "parameters": [
          {
            "name": "base",
            "in": "query",
            "description": "Base token address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quote",
            "in": "query",
            "description": "Quote token address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Historical chart window",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PriceHistoryPeriod"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_PairPriceHistory"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/wallets/{addr}/balances": {
      "get": {
        "tags": [
          "crate::http::app::balances"
        ],
        "summary": "Balances for `addr` over every catalog token. A malformed address → `400`.",
        "operationId": "balances",
        "parameters": [
          {
            "name": "addr",
            "in": "path",
            "description": "Wallet address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_List_TokenBalance"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Amount": {
        "type": "object",
        "description": "A token amount: `raw` base units (never a float), a human-formatted `display`, and `usd` (`None`\nuntil pricing is wired).",
        "required": [
          "raw",
          "display"
        ],
        "properties": {
          "display": {
            "type": "string"
          },
          "raw": {
            "type": "string"
          },
          "usd": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
...
      "PairPricePoint": {
        "type": "object",
        "description": "One timestamped pair midpoint, quoted as `quote` per one `base`.",
        "required": [
          "timestamp_ms",
          "price"
        ],
        "properties": {
          "price": {
            "type": "number",
            "format": "double"
          },
          "timestamp_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "volume_usd": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "USD-denominated market volume for the interval when the source provides it."
          }
        }
      },
...
      "PriceHistoryPeriod": {
        "type": "string",
        "description": "The history window requested by a price chart.",
        "enum": [
          "7d",
          "3m",
          "all"
        ]
      },
...
      "Response_List_TokenBalance": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "A page of a collection: the items plus an opaque `next_cursor` (absent on the last page) and an\noptional `total`. Carried inside the response envelope's `result`.",
            "required": [
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "description": "One token a wallet holds: its on-chain `balance` and the `pullable` amount a pull could take now\n(`min(balance, allowance→settlement)`).",
                  "required": [
                    "token",
                    "balance",
                    "pullable"
                  ],
                  "properties": {
                    "balance": {
                      "$ref": "#/components/schemas/Amount"
                    },
                    "pullable": {
                      "$ref": "#/components/schemas/Amount"
                    },
                    "token": {
                      "$ref": "#/components/schemas/Token"
                    }
                  }
                }
              },
              "next_cursor": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "total": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Response_PairPriceHistory": {
        "type": "object",
        "description": "The envelope wrapping every response. `status_code` sets the HTTP status (never serialized).",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "result": {
            "type": "object",
            "description": "Historical market data for one oriented pair.",
            "required": [
              "base",
              "quote",
              "period",
              "points"
            ],
            "properties": {
              "base": {
                "type": "string"
              },
              "period": {
                "$ref": "#/components/schemas/PriceHistoryPeriod"
              },
              "points": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PairPricePoint"
                }
              },
              "quote": {
                "type": "string"
              }
            }
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
      "Status": {
        "type": "string",
        "description": "Whether an API call succeeded.",
        "enum": [
          "Ok",
          "Error"
        ]
      },
...
      "Token": {
        "type": "object",
        "description": "A token's identity + display essentials — embedded wherever a response references a token\nwithout the full [`Asset`] (pool base/quote, trade flows, roster).",
        "required": [
          "address",
          "chain_id",
          "symbol",
          "decimals"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "decimals": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "symbol": {
            "type": "string"
          }
        }
      },
...
    }
  }
}